reexport libusb types from libusb1-sys
This change would make it possible to enumerate and select devices in
ways `libftdi` does not directly support, as well as interoperate with
`rusb` and/or other USB-related crates.
On the other hand, it may be possible to get really confusing errors if
`libusb1-sys` and `libftdi` get linked against two different `libusb`
versions. Detecting that correctly is not possible at the moment as the
current `libusb1-sys` versions do not export the required metadata.